requirements.txt
# ============================================================================
# KIVENDTOUT - DÉPENDANCES PYTHON
# ============================================================================

# Core
python-dotenv==1.0.0

# Base de données
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.1
motor==3.3.2

# Apache Airflow
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.16.0

# dbt
dbt-core==1.7.7
dbt-postgres==1.7.7

# Apache Kafka
kafka-python==2.0.2

# Apache Flink (PyFlink)
apache-flink==1.18.0

# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
polars==0.20.3

# MinIO (S3)
minio==7.2.3
boto3==1.34.24
s3fs==2024.2.0

# Machine Learning
tensorflow==2.15.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
scikit-learn==1.4.0
pillow==10.2.0
pytesseract==0.3.10

# Data Quality
great-expectations==0.18.8

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
rich==13.7.0

# Jupyter (pour notebooks ML)
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.28.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Data Generation (pour tests)
faker==22.2.0
